requirements.txt
chardet==3.0.4
pdfminer.six==20181108
